% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{group_lasso_gram}
\alias{group_lasso_gram}
\title{Solve group sparse regression}
\usage{
group_lasso_gram(
  G,
  g,
  Grp,
  lambda,
  beta0 = NULL,
  refine = T,
  max_iter = 1000,
  tolerance = 1e-06
)
}
\arguments{
\item{G}{The Gram matrix X'X / n}

\item{g}{The matrix X'y / n}

\item{Grp}{The group assignment matrix, indexes starting from 1}

\item{lambda}{The regularization parameter}

\item{beta0}{The initial value of beta}

\item{refine}{Whether to perform debiasing on the support}

\item{max_iter}{The maximum number of iterations}

\item{tolerance}{The convergence tolerance}
}
\value{
C The estimated coefficients matrix
}
\description{
Solve group sparse regression
}
\examples{
# Generate random data
set.seed(123)
n <- 100
p <- 20
X <- matrix(rnorm(n * p), n, p)
Y <- matrix(rnorm(n * p), n, p)
# Compute Gram matrix and cross-product
G <- crossprod(X) / n
g <- crossprod(X, Y) / n
# Define group assignments
Grp <- matrix(sample(1:5, p * p, replace = TRUE), nrow = p)
# Set regularization parameter
lambda <- 0.1
# Estimate matrix with group Lasso penalty
C_est <- group_lasso_gram(G, g, Grp, lambda)
}
